Move unused functions to attic.
